+++ lib/rubocop/options.rb
@@ -353,6 +353,7 @@ module RuboCop
                              '  [w]orst',
                              '  [t]ap',
                              '  [q]uiet',
+                             '  [a]utogenconf',
                              '  custom formatter class name'],
       out:                  ['Write output to a file instead of STDOUT.',
                              'This option applies to the previously',